refactor(UnionBan): refactoring UnionBan data processing logic
